+++ lib/src/constants.dart
@@ -1,2 +1,3 @@
 const shaLength = 32;
 const authenticationDataLength = 164;
+const signatureDataLength = 70;