+++ lib/src/models/get_assertion_options.dart
@@ -1,8 +1,9 @@
 import 'dart:typed_data';
 
 import 'package:json_annotation/json_annotation.dart';
-import 'package:webauthn/src/models/converters/uint8list_converter.dart';
-import 'package:webauthn/src/models/public_key_credential_descriptor.dart';
+
+import './converters/uint8list_converter.dart';
+import './public_key_credential_descriptor.dart';
 
 part 'generated/get_assertion_options.g.dart';