+++ lib/src/util/credential_safe.dart
@@ -7,6 +7,7 @@ import 'package:crypto_keys/crypto_keys.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:local_auth/local_auth.dart';
 
+import './webauthn_cryptography.dart';
 import '../db/credential.dart';
 import '../exceptions.dart';
 
@@ -32,8 +33,6 @@ class CredentialSafe {
   final CredentialSchema _credentialSchema;
   final LocalAuthentication _localAuth;
 
-  static final keyCurve = curves.p256;
-
   Future<bool> supportsUserVerification() async {
     if (authenticationRequired) {
       return await _localAuth.isDeviceSupported();
@@ -43,7 +42,7 @@ class CredentialSafe {
 
   /// Generate a new ES256 KeyPair and store it to our secure storage using the given alias
   Future<KeyPair> _generateNewES256KeyPair(String alias) async {
-    final keypair = KeyPair.generateEc(keyCurve);
+    final keypair = KeyPair.generateEc(WebauthnCrytography.keyCurve);
     final pk = keypair.privateKey as EcPrivateKey;
     final pub = keypair.publicKey as EcPublicKey;
 
@@ -69,12 +68,18 @@ class CredentialSafe {
     if (encoded != null) {
       final cborList = cbor.decode(base64.decode(encoded)) as CborList;
       final eccPrivateKey = cborList[0].toObject() as BigInt;
-      final pk = EcPrivateKey(eccPrivateKey: eccPrivateKey, curve: keyCurve);
+      final pk = EcPrivateKey(
+        eccPrivateKey: eccPrivateKey,
+        curve: WebauthnCrytography.keyCurve,
+      );
 
       final xCoordinate = cborList[1].toObject() as BigInt;
       final yCoordinate = cborList[2].toObject() as BigInt;
       final pub = EcPublicKey(
-          xCoordinate: xCoordinate, yCoordinate: yCoordinate, curve: keyCurve);
+        xCoordinate: xCoordinate,
+        yCoordinate: yCoordinate,
+        curve: WebauthnCrytography.keyCurve,
+      );
 
       return KeyPair(publicKey: pub, privateKey: pk);
     }
@@ -152,8 +157,10 @@ class CredentialSafe {
 
     final encoded = cbor.encode(CborMap({
       const CborSmallInt(1): const CborSmallInt(2), // kty: ECS key type
-      const CborSmallInt(3): const CborSmallInt(-7), // alg: ES256 sig algorithm
-      const CborSmallInt(-1): const CborSmallInt(1), // crv: P-256 curve
+      const CborSmallInt(3): const CborSmallInt(
+          WebauthnCrytography.signingAlgoId), // alg: ES256 sig algorithm
+      const CborSmallInt(-1): const CborSmallInt(
+          WebauthnCrytography.keyCurveId), // crv: P-256 curve
       const CborSmallInt(-2): CborBytes(xCoord), // x-coord
       const CborSmallInt(-3): CborBytes(yCoord), // y-coord
     }));